+++ change_log.md
@@ -6,7 +6,9 @@
 - edit/delete tag options now accessible via right click context menu.
 - tags can be marked as "excluded" (archived) which will exclude its craft from the results.
 - which tags are selected/excluded is now persisted over game reloads.
+- Added dropdown contect menu on craft in main list.
 - other UI functionality tweaks
+
 
 0.0.1 - first beta test release